Use regular 1.4 baksmali for Kitkat (for now)
The patches are failing to apply...I had heard of this before, but I thought
it was because the patches were made with 1.4. I made the patches with 2.0 and
used 2.0 in the auto-patcher---no dice. The 1.4 baksmali in the auto-patcher
had the same problem...every smali binary is producing output that rejects
patches from 2.0....hmmmmm.