Update for Android 4.4 (now the default)
This has provisions for the new AOSP placement of Mms (/system/priv-app) and
we now are using baksmali-2.0 for 4.4. As of right now, you need to get that
decompiler jar yourself.

This also adds a new user variable BAKSMALI_LOC, which is the location of the
decompiler.

Finally, the file structure and dir names of the holding folders has been
resolved. We are back to using the more sensible 'stock-framework' structure
that we have been using since jump.