Bundle baksmali and baksmali-2.0; only remove source patches if lock is present
This eases the discovery necessary for a new party, we now bundle all used
tools. Have also silenced an error if the API is not discoverable in the
case of no lock being found (meaning that no patches are applied to their build
system). In the case of no found lock, we no longer attempt to remove patches
from source code, which speeds up making patches in those cases.

some misc fixes. orderbuilds.sh is still a bit of a mess.